part2/requirements.txt
array_record
datasets
flax
gcsfs
google-cloud-storage
grain-nightly
hf_transfer
jax[tpu] -f https://storage.googleapis.com/jax-releases/libtpu_releases.html
orbax-checkpoint
optax
tpu-info
transformers